BE/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
google-genai>=0.8.0

ta>=0.11.0
